# gin-api-server/internal/swagger/swagger.json
{
    "schemes": [
        "http"
    ],
    "swagger": "2.0",
    "info": {
        "description": "funnydb api doc",
        "title": "FunnyDB",
        "contact": {},
        "version": "0.0.1"
    },
    "basePath": "/api/v1",
    "paths": {
...
        "/current/password": {
            "put": {
                "description": "修改当前用户密码",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "用户"
                ],
                "summary": "修改当前用户密码",
                "parameters": [
                    {
                        "description": "body",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/service.UserUpdatePasswordRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "{}",
                        "schema": {
                            "type": "object"
                        }
                    }
                }
            }
        },
        "/current/profile": {
            "put": {
                "description": "修改当前用户信息",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "用户"
                ],
                "summary": "修改当前用户信息",
                "parameters": [
                    {
                        "description": "body",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/service.UserUpdateProfileRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/model.UserDTO"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "model.UserDTO": {
            "type": "object",
            "properties": {
                "email": {
                    "type": "string",
                    "example": "demo@xmfunny.com"
                },
                "email_verified_at": {
                    "type": "integer",
                    "example": 1635696000000
                },
                "id": {
                    "type": "integer",
                    "example": 1
                },
                "is_admin": {
                    "type": "boolean",
                    "example": false
                },
                "last_active_at": {
                    "type": "integer",
                    "example": 0
                },
                "logo": {
                    "type": "string",
                    "example": " "
                },
                "name": {
                    "type": "string",
                    "example": "demo"
                }
            }
        },
...
        "service.UserUpdatePasswordRequest": {
            "type": "object",
            "required": [
                "new_password",
                "new_password_confirm",
                "old_password"
            ],
            "properties": {
                "new_password": {
                    "type": "string"
                },
                "new_password_confirm": {
                    "type": "string"
                },
                "old_password": {
                    "type": "string"
                }
            }
        },
        "service.UserUpdateProfileRequest": {
            "type": "object",
            "properties": {
                "name": {
                    "type": "string"
                }
            }
        }
    },
    "securityDefinitions": {
        "ApiKeyAuth": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}